[PATCH] Code emission for branches
This patch adds support for forward and backward branch code emission.  For
backward branches, we simply take the offset of the first instruction in the
target basic block.  Forward branches are added to a list of unresolved
branches in the target basic block.  The branch instructions are then
back-patched when the basic block is being emitted.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>