[PATCH] Simple parsing of INVOKEVIRTUAL
Add parsing of INVOKEVIRTUAL that stores the method index into a new
expression type that can be used to generate indirect calls.  The
current code looks up one of the target methods to determine method
signature which is probably not the right way to do it.  But that's
good enough for us now.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>