Fix operand ordering for binary operations in instruction selector
The subtraction acceptance test was failing because the operands
were in wrong order -- minuend was being subtracted from the
subtrahend. While the operands are in natural order in the expression
tree, the order for instructions is reverse as we follow the AT&T
convention where source comes first and destination second.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>