jit: fix dup bytecode parsing
The current code converts the following bytecode instruction sequence:

   6:   new     #3; //class java/lang/Object
   9:   dup
   10:  invokespecial   #4; //Method java/lang/Object."<init>":()V
   13:  astore_0

into a tree where there are two EXPR_NEW expressions. This is of course
wrong as the instance we actually initialize is different from the one
we store to local variable zero.

The regression tests still fail because we need instruction selection for
EXPR_VAR -> reg and STMT_STORE when the destination is EXPR_VAR.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>