jit: introduce struct register_info
This introduces struct register_info that will be used for live interval
splitting.  Each struct operand has one or two register info structs embedded
into them and the idea is that each will have a pointer to the struct
live_interval it belongs to so that when an interval is split, we can simply
change that pointer and code emission works as expected.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>