jit: don't spill or reload empty ranges
If liveness analysis doesn't find any instructions that are within a live
range, ->start is ~0UL and ->end is 0. So test for that in spill/reload and
skip those intervals. Although an empty range is unlikely to be spilled in real
life, at least I can now run the regression suite with every interval always
spilled.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>