jit: convert register allocator to follow algorithm by Wimmer et al
In preparation for live interval splitting and interval spilling, convert the
current register allocation algorithm that is based on [Poletto99] to follow
the version by [Wimmer05] described in:

  Wimmer, C. and Mössenböck, H. 2005. Optimized Interval Splitting in a Linear
  Scan Register Allocator.

The algorithm and related data structures are also explained in more detail in:

  Wimmer, C. 2004. Linear Scan Register Allocation for the Java HotSpot™ Client
  Compiler.

Note: this patch does _not_ implement splitting or spilling. Instead, assert()
is triggered if the register allocation reaches those states.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>