constant_pool: don't allocate entries separately
It doesn't really make sense to call malloc() for 2- and 4-byte objects,
so this patch changes the various constant pool entries to be stored
directly inside the constant_pool itself.

This makes the code much more readable, and actually reduces overall
memory usage too.

This also fixes a bug in the previous allocation code where an allocation
failure did not get reported.

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>