test: Enable MethodInvocationAndReturnTest in the regression suite
The commit 536d0fbcd2c247cf2d8b62c8281b334b040825da ("Removed EXPR_VAR in favor
of EXPR_TEMPORARY") fixed MethodInvocationAndReturnTest because we're now using
EBX for temporaries. The fix is not complete, of course, but no need to keep a
passing test disabled. We need the coverage anyway as Arthur continues his work
on the register allocator so we can get rid of the unnecessary get_fixed_var()
calls.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>