jit: replace ->insn_start with struct bytecode_buffer
As a preparation for mimic stack spilling, replace the use of ->insn_start
field in struct parse_context with a proper struct bytecode_buffer abstraction.
Now we are able to set up struct parse_context at the beginning of a basic
block rather than having to update ->insn_start for every bytecode.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>