jit: fix bug in register allocator when splitting intervals
The unhandled intervals list was browsed with list_for_each_safe, which
does not work if the immediate successor is modified. This happens when
splitting intervals and inserting the split child into the list - the
bug was that the child would be iterated over (ie. not treated), which
leads to unassigned registers.

Signed-off-by: Arthur HUILLET <arthur.huillet@free.fr>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>