jit: fix spill/reload emission to use correct machine register
Impact: fix generated code

The code used to spill and reload inside the register assigned to the
first interval (not the split child). As a result, if the split child
and the parent intervals were not allocated the same hardware register,
the generated code was wrong.

This fixes the long multiplication regression test, which now generates
correct machine code.

Signed-off-by: Arthur HUILLET <arthur.huillet@free.fr>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>