jit: sorting of basic blocks by their 'start' counter after convert_to_ir().
Control is transferred between basic blocks not only by branch
instructions such as goto/if*/athrow but it can be transferred between
adjacent basic blocks implicitly - after the last instruction of some
basic block has been executed, the control enters the following basic
block.

The simplest way to handle this case is to emit native code such that
basic blocks are in the same order as they are in bytecode. In that way,
if there was implicit control transfer between basic blocks in bytecode
there will also be one in jitted code. This is achieved by sorting basic
blocks by the 'start' counter before emission and after all basic blocks
are generated.

One of the impacts of this fix is correct compilation of for loops:

  public static void testForLoop() {
    int i;
    for (i = 0; i < 10; i++)
    assertEquals(i, 10);
  }

[ penberg@cs.helsinki.fi: fixes, cleanups ]
Reviewed-by: Arthur HUILLET <arthur.huillet@free.fr>
Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>