jit: fix interval splitting so it marks the child with REG_UNASSIGNED
When splitting intervals, the split child was given the same hardware
register as the parent.
This is incorrect because the split child has by definition not been
assigned any register at this point: REG_UNASSIGNED is the correct value
for ->reg.

This is a correctness fix: split childs with !REG_UNASSIGNED would make
it impossible for the register allocator to detect when
try_allocate_free_reg failed - and therefore assuming a reg had
successfully been allocated, whereas it should spill one instead.

Signed-off-by: Arthur HUILLET <arthur.huillet@free.fr>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>