jit: basic block splitting after athrow and return
Splitting after ATHROW and RETURN is required for basic blocks to be
generated at every exception handler entry so that
detect_exception_handlers() can find them all.

The correctness of the CFG generation for exception handlers is based on
the assumption that control can not be transferred to exception handler
basic block entry in other way than by catching an exception. This
implies that before every exception handler there must be one of the
following: return, athrow or a branching instruction.

[ penberg@cs.helsinki.fi: cleanups ]
Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>