jit: native_ptr_to_bytecode_offset() should deal with pointers not pointing directly at instruction beggining.
Upon exception throwing when it comes up that there is no handler in
current method, then we must unwind to the previous one. But we know
only return address and do not know the size of a call instruction that
was used. The solution is to decrement the return address pointer, and
then call native_ptr_to_bytecode_offset() which will return bytecode
offset of the call instruction because the given address belongs to that
instruction.

The solution utilises the fact, that basic blocks are emitted in order
they appear in the list and instructions are emitted in order they
appear in instruction lists.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>