x86: Fix overlapping return values
As explained by Arthur Huillet, if we use a fixed register ("eax" and
"edx") for storing the return value of a method, we can end up in two
overlapping ranges with the same fixed register.

Therefore, make sure we move the return value in "eax" and "edx" to a
non-fixed temporarily immediately after a CALL instruction. Doing so
fixes recursive fibonacci algorithm (and lots more).

Analyzed-by: Arthur Huillet <arthur.huillet@free.fr>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>