opcodes: generate opcode mnemonics using the JVM specification
Don't rely on Jam VM's definitions of the opcodes. Use a script that
downloads the opcode mnemonic definitions and generates a header file.

It probably doesn't matter much, but this is a way to avoid copying
the definitions provided by Jam VM, which we cannot do if we want to
completely get rid of Jam VM.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>