attribute_info: introduce new attribute_array type
We introduce a new type, attribute_array, which holds simply the size
of the array and a pointer to the array itself. This way, we can reuse
attribute the lookup function on this array (without passing too many
arguments around, since they are now collected in their own struct).

Also implement cafebabe_attribute_array_get(), which looks up the
first attribute with the given name.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>