jit, x86: spill exception object reference at exception handler entry
Impact: this fixes segmentation fault occuring when executing empty
exception handlers compiled with ecj. The ecj compiler generates a
'pop' instruction when exception handler is empty. This causes that
exception object reference is not popped from native stack when
exception handler is executed and that leads to native stack
corruption.

Solution is to always spill the reference to a well known location.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>