regression: Introduce and use VM.exit()
We don't want to use java.lang.Runtime because the class depends on
various Java APIs that Jato cannot cope with. It's not a problem now as
we use JamVM for initializing static fields but it will be when we
switch to cafebabe.

Unfortunately we cannot use VMRuntime.exit() directly because it's
package protected so therefore introduce an internal VM.exit() and use
that in the regression test suite.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>

[slurped this patch from mainline]
Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>