constant_pool: fix utf8 string comparison error
The utf8 string is not null-terminated, but comes with a length
parameter. Before this patch, we used strncmp() and limited the
comparison only the first part of the utf8 string. But this is wrong;
for example, these strings match: "foo" (utf8) and "foobar", because
the comparison stops after the three first characters (which are all
equal).

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>