jato: make stdout and stderr both unbuffered
This really helps debugging, for example when we want to run Valgrind
on jato and pipe the result to a pager. First of all, since we have mixed
stdout/stderr printouts, the lines will appear out of order. Then
valgrind comes along and prints to stderr, which means that we now have
three mixed sources of output. This makes everything strictly in-order.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>