jit, x86: support for throwing NullPointerException from SIGSEGV handler
This significantly reduces code required for doing null checks. In
some places (eg. invokevirtual) there is no need to insert any extra
code for null checks to work because the reference to be checked is
being dereferenced by native code and will trigger SIGSEGV when it is
null. In other places null check can be reduced to a single
instruction like "test (%eax),%eax".

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>