jit: mimic stack spilling and reloading
There's a bug in how we handle the mimic stack on basic block boundaries. The
problem is that a mimic stack can be non-empty at the end of a basic block
because of, for example, the ternary operator. You can find a full explanation
of the problem in the paper Fast, effective code generation in a just-in-time
Java compiler by Parikh et al (1998).

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>