jit, x86: always push exception object reference on stack
Throwing code will always push exception reference because exit basic block
will always safely discard it. This also fixes a crash when running jato on the
following code:

  public class Test {
    public static void main(String[] args) {
        throw new RuntimeException();
    }
  }

That's because code selected for STMT_ATHROW didn't handle the case where the
jump target was NULL which meant that a jump to exit block should be made and
no exception object should be pushed.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>