scripts: allow file names in tests' OBJS list to have the same name
It was impossible to link test which had in its OBJS list files with same
name, eg. ../../arch/x86/exception.c and ../../jit/exception.c.
The reason for that is that only basename of target file is taken
so that both files resolve to exception.o. Solution for this problem
is to generate names for objects using whole path information, so
that the first one will resolve to p-p-arch-x86-exception.o and the
second one to p-p-jit-exception.o

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>