jit, vm: mechanism for signaling exceptions from jato functions
It is not possible to throw exceptions directly from jato functions
because we cannot unwind C functions. Instead we store the exception
pointer in current execution environment and set "exception_guard"
pointer to a hidden guard page. When control will be returned to jit
method the code will poll memory pointed by "exception_guard". When
the pointer points to hidden page it will trigger SIGSEGV. The signal
handler takes care of throwing the exception.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>