vm: back-end support for java.lang.Object.getClass()
This patch adds an ->object member to struct vm_class, which points to
an instance of java.lang.Class which represents the vm_class.

It's all a little troublesome. Especially the nasty recursion we get
if we're not careful. java.lang.Object may want to load java.lang.Class
in order to initialise its ->object, and java.lang.Class may want to
load java.lang.Object because that is its superclass. What I did to
solve this was to split the ->object initialisation into a separate
function and preload both java.lang.Object and java.lang.Class before
calling this extra initialisation function.

It's a bit ugly. I'd like to solve this in a cleaner way (or at least
define the rules clearly, once and for all). The JVM spec probably has
something to say on the matter.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>