vm: provide vm_class_is_assignable_from()
Following some discussion with Tomasz Grabiec on IRC, I decided to
introduce this function, which checks whether a cast from one class to
another is possible.

The "is instance of" name doesn't really fit here, because we are
talking about two classes, neither of which is an instance of the other.

Objects can be instances of classes, however, and this function, called
vm_object_is_instance_of() takes an object and a class as arguments.

Reference: http://folk.uio.no/vegardno/jato-irc-logs/2009-06-20.txt
Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>