jit: remove is_jit_method() and introduce is_native().
The is_jit_method() function does not perform the check which
it's name suggests. It checks whether address is above text segment
which is wrong because it clasifies as JIT methods functions which
are not JIT: trampolines, linked library functions, kernel functions.

In place of is_jit_method() a new function is introduced - is_native().
It checks whether given address is located outside the _heap_. It
returns true to all functions but JIT methods and trampolines. It can
be used in place of is_jit_method() where it is reasonable to assume
that the address can not be in a trampoline.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>