jato: improve class preloader
We use the preloader to automatically preload a few select classes
which are essential to VM operation. If any of these fail, we cannot
start the VM at all.

This allows us to easily reference these classes from the VM core
without having to go through the class loader (which could in fact
fail to load the classes) -- that saves us a bit of error handling
as well (since we can simply assume that the loading was fine).

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>