vm: the foundation for java stack trace collecting
A number of utility functions are introduced for backtrace traversal
to assist java stack trace collecting. Only JIT methods and VM natives
should appear in stack trace so all native functions and trampolines
must be filtered out.

A new per-thread variable bottom_stack_frame is introduced to detect
when to stop tracig back. It is set to a native frame pointer of the
function that calls the main() method.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>