insn-selector: don't use fixed var for EXPR_ARRAYLENGTH
When compiling the expression (x % array.length), the "array.length"
would put the result in %eax. The "idiv" instruction (used by the %
operator) takes a fixed input in %eax and one additional operand; the
result is that we were outputting "idiv %eax", which is nonsensical (and
wrong) because it divides %eax by itself.

Acked-by: Arthur HUILLET <arthur.huillet@free.fr>
Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>