jit: mmap 256 MB buffer for jit text
The current on-demand JIT text allocation causes the buffer to be
relocated once a page has been filled. This means that all branch and
direct call targets are invalidated. Fix that up by mmapping a
contiguous 256 MB buffer and making sure only one thread emits code at a
time.

Thanks to Arthur and Vegard for debugging this!

Analyzed-by: Vegard Nossum <vegard.nossum@gmail.com>
Analyzed-by: Arthur HUILLET <arthur.huillet@free.fr>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>