jit: fix bug in convert_dup*() function family
It is incorrect to push the same instance of temporary expression twice onto
mimic-stack because some expressions may use that temporary to store the result
and thus unexpectedly (and incorrectly) change the input of another expression.
Below is a bytecode sequence that exposes the bug:

    iconst_2
    dup

    iconst_1
    iadd
    istore_1

    iconst_2
    iadd
    istore_2

If dup pushes the same temporary expression twice then the same register will
be used as left operand (and the result also) for both ADD operations. This
will cause that the result of the first ADD operation will be stored to the
register which is input to the second ADD operation.

Solution for this is to push different temporaries onto mimic-stack
holding the same initial value.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>