jit: fix bug in convert_getstatic() and convert_getfield()
We should not push EXPR_CLASS_FIELD and EXPR_INSTANCE_FIELD onto
mimic-stack because another operation can change the value of the
field and when EXPR_*_FIELD will be popped from stack it's value will
be different from the one that it had when it was pushed. It breaks
language semantics.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>