jit: temporaries for mimic stack spilling are now recursively propagated
This fixes potential bugs when "master-master" relationships:

  A     C     E
   \   / \   / \
    \ /   \ /   \
     B     D     F

A and F are slaves, the rest are masters.  If we browse blocks in order, it is
B that is going to pick the temporary.

We easily see that this temporary must be used by *all* the blocks for their
mimic stack spill slots.

Recursive propagation of the temporary choice from B into C, D, E does this.

Signed-off-by: Arthur HUILLET <arthur.huillet@free.fr>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>