jit: fix method_real_argument_count
We're scanning the method signature to determine the number of "real"
arguments to pass for an "invoke" opcode. We use the function
method_real_argument_count() for this. But it doesn't skip class names,
so if there is an uppercase J or D in the class name (of the argument
type), we count one argument less than we should and push one argument
less on the stack.

Reviewed-by Arthur HUILLET <arthur.huillet@free.fr>
Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>