vm: initialize static fields (completely) to null
We need to initialize static fields to zero even before the class
initializer runs (since the initializer may refer to static fields).

This still needs a lot of work, but here is a preliminary fix that
initializes a whole pointer instead of just a single byte. We are
still not using the constant (as found in the ConstantValue
attribute), but this gets us past java.lang.Runtime initialization.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>