insn-selector: emit traps for static final fields
A sentence in the JVM spec gave me the impression that accessing a final
class field should not trigger class initialization, but this appears to
be wrong. It appears that this sentence was intended for the compiler,
since GNU classpath's System.out is in fact static final and without this
patch, we crash (with a NullPointerException) on trying to call
System.out.println because System does not initialize when we access
System.out.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>