vm: introduce vm_class_is_equal_to() to compare classes
Classes can be equal even if the pointers are not equal.
That's because array classes are not cached and the same class may
be represented by different copies of struct vm_class.

The proper soultion would be to cache array types also but it is
not something that can be easily done now.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>