x86: add emit_really_indirect_jump_reg()
It turns out that the existing emit_indirect_jump_reg() isn't what it
claims to be -- it emits a _direct_ jump to the address specified in
the register.

Compare:

   ff 20     jmp    *(%eax)	# indirect
   ff e0     jmp    *%eax	# direct

A better fix would of course be to call them emit_direct_jump_reg()
and emit_indirect_jump_reg(). (But for my convenience, I propose that
as a delta patch once invokeinterface has been merged.)

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>