jato: switch license
Seems that we forgot to switch this one when importing the cafebabe
files. Note: The file says my name, mostly because this file was
originally forked from the cafebabe example program. But I welcome
(and encourage) the others to add their names as well; this is the
main program source file, and it gives the impression that I am the
sole author of the whole program, which is obviously plain wrong.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>