x86: fix bug in invokeinterface and invokevirtual args cleanup
method->args_count should be used to get the size of arguments,
not nr_args(args_list). The former is the right thing to do here as
vm_method_init() calls count_arguments() to initialize it which counts
the number of 32-bit memory locations used by the arguments.

This bug led to stack overflow with the following code:

public class Test {
    public void c(long x) {
    }

    public static void main(String[] args) {
        Test t = new Test();
        int x = 0;

        while (x == 0) t.c(1l);
    }
}

[ penberg@cs.helsinki.fi: fix up compilation warnings ]
Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>