vm: remove trampolines for JNI methods
This introduces a better approach for calling JNI methods. The
approach is based on assumption that if native method is not a VM
native then it is a JNI method. This lets us to prepare JNI call
arguments at compilation time.

When method is prepared in vm_method_prepare_jit() it is checked
whether native method is one of VM native methods and the address of
it is stored in the ->vm_native_ptr field.

When converting method invokation it is checked whether method is a
JNI native method. If this is the case then JNI environment pointer is
appended to call arguments. When static JNI method is called the
class->object pointer is also appended according to the JNI
specification.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>