Fix use-after-free in fixup_static_at()
Ensuring that the class is initialized may cause fixup to be called
again, and when we return to the first fixup call, the fixup list will
be invalid. The solution is to not access the list anymore after we've
ensured that the class has been initialized.

This was introduced in:

Author: Vegard Nossum <vegard.nossum@gmail.com>
Date:   Tue Jul 7 15:36:28 2009 +0200

    vm: fix deadlock (error in invokestatic handling)

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>