jit: introduce subroutine inlining
Inlining of subroutines eliminates JSR, JSR_W and RET instructions
from bytecode. These instructions are generated by java compilers to
handle try-finally blocks. As the effect of inlining the body of
each subroutine is copied in place of every JSR instruction which
calls that subroutine. Inlining modifies method's code, exception
table and line number table.

The algorithm is based on the following work:
 "Subroutine Inlining and Bytecode Abstraction to Simplify Static
  and Dynamic Analysis", Cyrille Artho, Armin Biere.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>