vm: fix argument passing for JNI methods
The ->args_count is now correctly set for JNI methods
to include the JNI environment pointer (for all JNI methods)
and a class object pointer (for static JNI methods).

Argument passing for JNI methods is now done in convert_args()
to let JNI arguments to be passed in registers like for
other methods.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>