vm: introduce alloc_offset_guard()
Function allocates a memory region with two sub-regions. The first one
contains addresses from range [0:valid_size[ with access allowed. The
second region contains addresses from range [valid_size:overflow_size[
and access to this region triggers SIGSEGV.

This mechanism is useful for checks which fail very rarely and need
to be as fast as possible.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>