Move the mmix architecture to arch/mmix
Mmix is an architecture that is not actually used on any real hardware
but we do use it for "architecture independent" JIT tests. As people
keep breaking "make test" constantly when changing architecture specific
code, lets move the damn thing under "arch" proper so people know it
needs to be maintained.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>