vm: fix call stack traversal when called from JNI trampoline
When JNI method linking fails an instance of
java.lang.UnsatisfiedLinkError is thrown from JNI trampoline
(jit_native_trampoline()). When stack trace is created for this
exception, stack walker is trying to use the
method->compilation_unit->native_ptr to get the JNI method's address,
but this pointer is not valid because linking failed. To solve this
problem we add a new field to struct stack_trace_elem to hold a
pointer to compilation unit. For JNI stack trace elements this field
is used and no jit_lookup_cu() is performed on ->addr. This also makes
stack traversal faster.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>