jit: Control flow resolution
The linear scan register allocator simplifies the control flow graph to
a linear list of basic blocks. However, this is not sufficient when
intervals are split which means we need an additional resolution step in
the allocator.

The algorithm is described in Section 5.8 ("Resolving the Data Flow)" of
"Linear Scan Register Allocation for the Java Hotspot Client Compiler"
by Christian Wimmer.

[ penberg@cs.helsinki.fi: cleanups ]
Signed-off-by: Arthur Huillet <arthur.huillet@free.fr>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>