jit: fix bb_last_insn so it returns the last LIR instruction of the block
When a live interval ends at a basic block boundary and is spilled, the
spill instruction will appear after the final jump of the basic block.
In that case, bb_last_insn really ought to return the struct insn
corresponding to the jump, not the spill instructions that have been
added afterwards.
This is a bugfix.

Signed-off-by: Arthur HUILLET <arthur.huillet@free.fr>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>