jit: make trace output thread safe.
Each thread has its own trace buffer to which strings are printed with
trace_pritnf(). When trace_flush() is called the buffer's content is
printed to stderr. trace_flush() is synchronized with other threads.

if -Xtrace:threads is passed to jato, every line of trace will be
printed with trailing current thread's name.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>