vm: fix in classloader_load()
When a class was loaded with name in a dot form and if another time
classloader would be asked about the same class but in a slash form
then classloader will try to load the class again. We should convert
class names to slash names at classloader_load() entry to avoid
loading the same class twice.

This bug led to infinite loop of initialization of
"gnu/java/awt/peer/gtk/GtkComponentPeer" because it is referenced from
somewhere in its initialization code with a dot form name:
"gnu.java.awt.peer.gtk.GtkComponentPeer".

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>