jit: do not connect athrow with exception handlers
There is no need to create CFG edges after athrow to exception
handlers because VM stack is empty at exception handler
entry. Exceptional flow is special because it "resets" the mimic stack
and we do not need to resolve anything across this boundary yet. We do
not create any edges for every PEI (eg. idiv, invokevirtual), so we
should not confuse the reader with creating adges after athrow.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>