jit: introduce EXPR_ARG_THIS for holding invocation 'this' pointer
The 'this' pointer is not always the first call argument. For JNI
function invocations it is the second argument. EXPR_ARG_THIS is
introduced to mark the argument holding object reference so that
the register which is assigned to it can be easily propagated up
during instruction selection.

Before that JNI invocations via invokevirtual were ignored which
leads to incorrect behavior.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>