vm: fix trace_flush()
We should not use strtok() to split string into lines because it
swollows consequtive '\n' characters.

We also do not flush the buffer tail not ended by new line character
because we would print the thread name marker ([main]) in the middle
of the line next time the buffer is flushed. Tracing buffer behaves
now as if line-buffered.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>