jit: optimize __analyze_use_def()
Now that use-def rules are corrected, lets optimize __analyze_use_def() a bit.
This reduces the time to run S.o.println() from 12 seconds down to 7 seconds. A
positive side-effect of this patch is that liveness analysis produces more
accurate results now which should reduce register pressure and the amount of
spills.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>