x86: fix INT -> BYTE conversion.
We are using movsbl instruction to sign-extend the byte value.
This instruction uses r/m8 addressing for source. There is no
equivalent for lowest byte of %esi and %edi in this addressing,
codes for those register are representing %dh and %bh registers.

This bug caused that CharsetEncoder.encode() did not work properly
preventing System.out.println() from correct behavior.

[ penberg@cs.helsinki.fi: cleanups ]
Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>