x86: implement double precision arithmetic
This adds support for most of double precision arithmetic. The only
missing things are conversions which will be added in the following
patches.

Operations on doubles are made using SSE2 instructions.  We want to
emulate those operations when current system is not equiped with SSE2
capable processor.  This needs fixing.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>