jit: alloc temporary with correct register type in convert_array_load()
We should really care which vm_type we assign to registers. Not only
it matters because FPU operations can be only done on SSE registers,
but also in some situations the only way to get the result type is to
query reg->vm_type (see OP_CMPG and OP_CMPL instruction selectors).

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>