jit: Fix empty interval spill bug
The bug resulted in attempt to spill an interval which
has no register assigned.

Initial interval for vreg 20 has use positions at 4 and 29.
Register is available only for positions 4-11 which results
in interval splitting. There are no free registers available
for the new interval (11-29) so it is split at next use position,
which results in new interval 24-29. This interval is marked
as needing to be reloaded. The parent interval however has
no register assigned.

20 (pos:  4-11):	ECX	non-fixed	spill (5)	no reload
20 (pos: 11-24):	-	non-fixed	spill (6)	reload (5)
20 (pos: 24-29):	ESI	non-fixed	no spill  	reload (6)

Proposed solution for this problem is to not generate intervals
which do not start with use position in try_to_allocate_free_reg().
In this example, we throw away the interval 11-24 will be dropped.
We can do that because register is not used in this range.
After this fix, we will have the following regalloc:

20 (pos:  4-11):	ECX	non-fixed	spill (5)	no reload
20 (pos: 24-29):	ESI	non-fixed	no spill  	reload (5)

Acked-by: Arthur HUILLET <arthur.huillet@free.fr>
Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>