jit: implement lookupswitch
Lookupswitch is implemented in the simmilar way tableswitch is.
Lookupswitch requires two basic blocks. The first one performs binary
search to find the branch target. It's ended by if statement which
checks if the result is NULL (target not found). In this case default
target is taken. Otherwise the second basic block is entered which
simply jumps to the found target.

Currently, binary search is implemented by a call to standard
bsearch() function. It can be optimized in the future by inlining the
binary search code for small lookupswitch tables.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>