jit: introduce per-method JNI trampolines.
We can't decide at compile-time whether called method is JNI.
That's because methods declared as native in one class can
be overriden in subclass and be non-native.

Solution for this is to call potential JNI methods like any
other methods. Method pointers for actual JNI methods
point to a trampoline which takes care of pushing
pointer to JNI environment and class object reference
for static methods.

Signed-off-by: Tomek Grabiec <tgrabiec@gmail.com>
Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>